Move rebasing tests to RebaseDateTimeSuite
